Added support for converters that create arbitrary content that write it to the file itself.
This commit adds the interface WritingConverter.
This mimics an Asciidoctor Ruby converter that includes the Asciidoctor::Writer module.
That is the converter writes the content itself to the target file.

This is also the precondition for allowing a converter to create non-stringy content.
Therefore the Converter interface is now a generic interface that has to declare the results it produces.

The Asciidoctor interface is extended with several generic convert & convertFile methods that take an expected result type so that the caller can receive arbitrary object content instead of only strings.
The existing convert methods delegate to these new methods with a expected result type of String.
If the result type does not match the actual result a ClassCastException will be thrown.