prevent multiple upload of artifacts to bintray
The bintray plugin allows two ways to define the artifacts to be uploaded.

* Maven Publications
* Configurations

We use both, as it is not possible to add the signatures for the jars to the maven publication.

(Or maybe it is possible, but I couldn't figure out how. Adding the signed artifacts to the publication like adding the
pom signature failed with an error while publishing to maven local.)

The _archives_ configuration holds the jar-artifacts and the signatures. And that is why the jar artifacts are doubled
and the signatures not.
Defining the _signatures_ configuration is the solution for every non distribution module.

One thing I was not able to prevent from uploading twice is the pom file.
But I think that is tolerable.

(Maybe it is not supposed to mix those two definitions and each of them is adding the pom transparently...)