Update version of asciidoctorj
Variable is adjusted in both README and antora config

Update jRuby version

It's the version that was used in the 2.5.7 release, NOT from the
most recent commit.

Update links, remove direct links to files

Add variable for group id

Use group-id variable for URLs

Add working links to direct files

HTTPS is required for the API calls to work.